Move character set related settings to it's own file
